[IMP] stiloconcepto: sync line names between PO and SO
This commit introduces an override of the `write` method on
`sale.order.line` to automatically update the 'name' field of related
purchase order lines when a Sale Order line is modified. The update only
applies to lines associated with products under the same section of the
SO, ensuring that multiple sections with the same product are correctly
distinguished.